move setup_udev_rules from utils to version to avoid importing heavy utils.py during install
